fix reporting on multi-clause callbacks
In case of multi-clause callbacks a missing spec error was raised for
clause bodies after the first @impl annotation. Fix the bug by
checking if we already have parsed a function/arity body with an
implementation annotation and pass this implementation to all subsequent
clauses.

Add a custom behaviour unit test.