Do not report missing docs for exceptions (#36)
* Fix cli args parsing for doctor.explain

* Option for requiring moduledoc on exceptions

In case of a `defexception` the convention is to not have a `moduledoc`.
A config option `exceptions_moduledoc_required` is added which by
default is set to true. If it is set to `false` then exceptions without
moduledocs are considered valid.

Exception struct type spec can be ignored, since the `__struct__` is
injected from the `defexception` macro. In case of an exception no
matter what the value of `struct_type_spec_required` is `N/A` will be
returned.

Include unit tests.