Properly collect functions for nested modules (#47)
There were two issues:
- The parent of a nested module would include all of the functions from its nested modules

- A nested module wouldn't include any of its functions

The first issue was caused because `ModuleInformation.load_user_defined_functions` wasn't taking into account nested `defmodule` statements.

To fix this:
- Add a `nesting_level` to the accumulator when looking for functions.
- Switch to using `Macro.traverse/4` so that we can properly keep track of the current module nesting level.
- Ignore any `def` statements at nesting level > 1.

The second issue was caused because the `modules` map in `Module.load_user_defined_functions` was not using fully-qualified names for the nested modules. As a result, when attempting to find the nested module's functions (which uses the fully-qualified name), the module's AST wasn't found.

To fix this:
- Convert the accumulator to include a module `stack`.
- Switch to using `Macro.traverse/4` so that we can properly maintain the stack as we enter and exit modules.
- Compute the fully-qualified name of any nested module and use that as a key in the `modules` map.

Note that nested modules can also include `defimpl`s, so we also check for those when looking for functions.